Remove setting the Content-Type when storing a file in object storage
Attachments were not displaying inline with Google Cloud Storage and
the Content-Type could not be overriden if this type was set.
Since `application/octet-stream` basically means the data is unknown
and Google/Amazon will default to an equivalent MIME type, there's
no sense it setting it.

This change has to be made in conjunction with https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/21265